requirements.txt
google-generativeai==0.4.1
pillow==9.5.0
streamlit==1.32.0
streamlit-option-menu==0.3.12